requirements.txt
websocket-client==0.35.0
python-vlc==1.1.2
soundcloud==0.5.0
